\name{add_grid}
\alias{add_grid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Add grid to a plot
}
\description{
  
  Add grid to a plot using abline.
  
}

\usage{
add_grid(at.y, at.x, col = "lightgray", lty = "dotted", lwd = par("lwd"))
}

\arguments{
  \item{at.x}{x axis grid values}
  \item{at.y}{y axis grid values}
  \item{col}{grid color}
  \item{lty}{line type}
  \item{lwd}{line width}
}

% \details{
  

% }

\value{
  The function adds a grid to a plot using abline.

%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ add }
\keyword{ grid }
\keyword{ plot }

