% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blockrand2randlist.R
\name{blockrand2randlist}
\alias{blockrand2randlist}
\title{Create a randomization list from a blockrand data.frame output}
\usage{
blockrand2randlist(x, file = NULL, projectTitle = "")
}
\arguments{
\item{x}{blockrand output}

\item{file}{path to file to save in (overwriting the contents). If NULL
the list is displayed.}

\item{projectTitle}{a character used in the page footer}
}
\description{
Create a randomization list from a blockrand data.frame output
}
\examples{
\dontrun{
library(blockrand)
set.seed(     6503529)
sample.size <- 140
Levels <- c("0","1")
block.size <- 10
randlist <- blockrand(n = sample.size,
                      num.levels = length(Levels),
                      levels = Levels,
                      block.sizes = rep(block.size/2, 2)
                      )
blockrand2randlist(randlist,
                   projectTitle = "PI - Project",
                   file = "randlist.xlsx" )

}
}

