% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/censorAt.R
\name{censorAt}
\alias{censorAt}
\title{Censor time to event end point at a certain time.}
\usage{
censorAt(time = NULL, status = NULL, censor.time = NULL)
}
\arguments{
\item{time}{Progression Indicator}

\item{status}{Progression Date}

\item{censor.time}{Death indicator}
}
\value{
A data frame to be used with \code{\link{cbind}}
}
\description{
Censor time to event end point at a certain time.
}
\examples{
time   <- c(100,150)
    status <- c(  0,  1)
    censor.time <- 120
    (original <- data.frame(time, status))
    censorAt(time=time, status=status, censor.time = 160)
    censorAt(time=time, status=status, censor.time = 150)
    censorAt(time=time, status=status, censor.time = 125)
    censorAt(time=time, status=status, censor.time = 75)

    time   <- c(100,150)
    status <- c(  1,  0)
    censor.time <- 120
    (original <- data.frame(time, status))
    censorAt(time=time, status=status, censor.time = 160)
    censorAt(time=time, status=status, censor.time = 150)
    censorAt(time=time, status=status, censor.time = 125)
    censorAt(time=time, status=status, censor.time = 75)
}

