% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chi_or_fisher.R
\name{chi_or_fisher_test}
\alias{chi_or_fisher_test}
\title{perform Chi-square or Fisher test as appropriate}
\usage{
chi_or_fisher_test(x, y)
}
\arguments{
\item{x}{first variable}

\item{y}{second variable}
}
\description{
perform Chi-square or Fisher test as appropriate. The function uses
\code{\link{fisherNeeded}} to decide whether to perform chi or fisher
test.
}

