% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/commaChar2Numeric.R
\name{commaChar2Numeric}
\alias{commaChar2Numeric}
\title{Convert a char with comma to a numeric}
\usage{
commaChar2Numeric(x)
}
\arguments{
\item{x}{vector of chars with commas}
}
\value{
a vector of numeric
}
\description{
Convert a char with comma to a numeric
}
\examples{
commaChar2Numeric(c("2,456", "1,234"))
}

