% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Comment}
\alias{Comment}
\title{Print a message inside an ascii border}
\usage{
Comment(..., border = "-", wrap = options()$width)
}
\arguments{
\item{...}{character vectors (pasted) forming the message to
be printed}

\item{border}{character used for header printing}

\item{wrap}{where to wrap}
}
\description{
Print a message inside an ascii border; useful for working
with \code{\link{sink}}, to leave messages in the output.
}

