% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cramerV}
\alias{cramerV}
\title{Cramer's V calculator.}
\usage{
cramerV(x)
}
\arguments{
\item{x}{an R object. Currently implemented only for tables.}
}
\value{
A scalar with Cramer's V
}
\description{
Cramer's V calculator.
}
\examples{
a <- rnorm(1000) > 0.2
b <- gl(2,500)
cramerV(table(a,b))
}
\keyword{Cramer}
\keyword{Phi}
\keyword{V}

