% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datePlus.R
\name{datePlus}
\alias{datePlus}
\title{Add a period of time to a Date object}
\usage{
datePlus(x, years = 0, months = 0, days = 0)
}
\arguments{
\item{x}{starting Date}

\item{years}{years to be added}

\item{months}{months to be added}

\item{days}{days to be added}
}
\description{
Add a period of time to a Date object
}
\examples{
(today <- Sys.Date())
datePlus(today, years = 1, months = 1, days = 1)
}

