% Generated by roxygen2 (4.0.2): do not edit by hand
\name{daysToMonths}
\alias{daysToMonths}
\title{Trivial function that converts number of days in number of months.}
\usage{
daysToMonths(days)
}
\arguments{
\item{days}{Numeric vector of days}
}
\value{
A numeric vector of months.
}
\description{
Trivial function that converts number of days in number of months.
}
\examples{
days <- 1:30
daysToMonths(days)
}
\keyword{day}
\keyword{days}
\keyword{month}
\keyword{months}

