% Generated by roxygen2 (4.0.2): do not edit by hand
\name{daysToYears}
\alias{daysToYears}
\title{Trivial function that converts number of days in number of years.}
\usage{
daysToYears(days)
}
\arguments{
\item{days}{Numeric vector of days}
}
\value{
A numeric vector of years.
}
\description{
Trivial function that converts number of days in number of years.
}
\examples{
days <- seq(1, 365.25*3, by = 60)
daysToYears(days)
}
\keyword{day}
\keyword{days}
\keyword{year}
\keyword{years}

