% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eng_maxima}
\alias{eng_maxima}
\title{A maxima engine for knitr}
\usage{
eng_maxima(options)
}
\arguments{
\item{options}{knitr options}
}
\description{
A maxima engine for knitr
}
\examples{
\dontrun{
## Add this to your document
library(knitr)
knit_engines$set(maxima = yapomif::eng_maxima)
opts_chunk$set("engine" = "maxima")
}
}

