\name{enSummary}
\alias{enSummary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  summary with sd
}
\description{
  
  This is a wrapper around summary which adds sd by default.
  
}

\usage{
enSummary(var)
}

\arguments{
  \item{var}{Vector to be passed to summary.}
}

% \details{
  

% }

\value{
  The function return same results of summary adding sd.

%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}

}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ summary }

