% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fisherNeeded.R
\name{fisherNeeded}
\alias{fisherNeeded}
\title{Is fisher test needed}
\usage{
fisherNeeded(x, y, threshold = 5, fullReport = FALSE)
}
\arguments{
\item{x}{the first variable}

\item{y}{the second variable}

\item{threshold}{cutoff value under indipendence}

\item{fullReport}{return a full report or only the \code{TRUE}/\code{FALSE} answer}
}
\description{
Is fisher test needed?
}

