% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fupCompleteness}
\alias{fupCompleteness}
\title{Calculate C index of of follow up completenes}
\usage{
fupCompleteness(time = NULL, status = NULL, cutoff = NULL)
}
\arguments{
\item{time}{Follow up (in days?)}

\item{status}{event indicator}

\item{cutoff}{(in days?)}
}
\value{
A list with individual follow up completeness and the
C index
}
\description{
Calculate C index of of follow up completenes.
}
\examples{
time   <- c(180, 12, 240, 250 )
status <- c(  0,  1,   0, 1   )

## example:
## quantify fup completeness to 200 days (eg minimum potential
## follow up in a hypotethic prospective trial)

fupCompleteness(time = time, status = status, cutoff = 200)
}
\references{
Clark T., Altman D., De Stavola B. (2002),
Quantification of the completeness of follow-up. Lancet 2002;
359: 1309-10
}
\keyword{completeness}
\keyword{event}
\keyword{follow}
\keyword{time}
\keyword{up}

