% Generated by roxygen2 (4.0.2): do not edit by hand
\name{groupProgID}
\alias{groupProgID}
\title{Group progressive id creator from unordered id vector}
\usage{
groupProgID(group)
}
\arguments{
\item{group}{a group vector}
}
\description{
Starting from a vector of group id (not necessarily ordered),
this function creates a progressive id inside each group.
}
\note{
To be implemented in C/C++
}
\examples{
set.seed(1)
## One factor id
x <- sample(gl(2,10, labels = c("A","B")))
cbind(x, groupProgID(x) )
## Two factors id
y <- sample(gl(2,10, labels = c("C","D")))
cbind(x, y, groupProgID(interaction(x,y)) )
}
\keyword{group}
\keyword{id}

