% Generated by roxygen2 (4.0.2): do not edit by hand
\name{groupProgID2}
\alias{groupProgID2}
\title{Group progressive id creator from ordered id vector}
\usage{
groupProgID2(group)
}
\arguments{
\item{group}{a group vector}
}
\description{
Starting from a vector of group id (that have to be ordered,
aka multiple elements from the same group must be in adjacent
rows, the group label is irrilevant), this function creates a
progressive id inside each group.
}
\examples{
set.seed(1)
## One factor id
x <- c("A", rep("C",3), rep("B",2))
cbind(x, groupProgID2(x) )
## example of sorting
db <- data.frame("group" = sample(gl(2,5)), "b" = Sys.Date() + 1:10)
(db <- db[order(- as.numeric(db$group), db$b),])
cbind(db, "seq" = groupProgID2(db$group))
}
\keyword{group}
\keyword{id}

