% Generated by roxygen2 (4.0.1): do not edit by hand
\name{header}
\alias{header}
\title{Print a message inside an header}
\usage{
header(msg = "", char = "-")
}
\arguments{
\item{msg}{message to be printed}

\item{char}{character used for header printing}
}
\description{
Print a message inside an header
}

