\name{inches}
\alias{inches}
\title{Translates from cm (centimeters) to inches.}
\usage{
inches(cm)
}
\arguments{
  \item{cm}{Numeric vector of centimeter.}
}
\value{
A vector with inches.

%% ~Describe the value returned %% If it is a LIST, use %%
\item{comp1 }{Description of 'comp1'} %% \item{comp2
}{Description of 'comp2'}
}
\description{
Translates from cm (centimeters) to inches.
}
\keyword{cm}
\keyword{inches}

