% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePath.R
\name{makePath}
\alias{makePath}
\title{Path-function generator.}
\usage{
makePath(path)
}
\arguments{
\item{path}{Base path of interest.}
}
\value{
A function that paste the directory set up in the call to
\code{makePath} to the argument passed in the current call.
}
\description{
This function creates another function that paste the directory set up in
the call to \code{makePath} to the argument passed in the current call,
allowing better directory parametrazion, auto-explaining code and less
typing. See example for usage.
}
\examples{

res.dir <- makePath("results")
tmp.dir <- makePath("/tmp")

## Eg for use with write.csv() or pdf()
res.dir("my_results.csv")
tmp.dir("graph.pdf")


}

