% Generated by roxygen2 (4.0.2): do not edit by hand
\name{NAreport}
\alias{NAreport}
\title{Report NA variables per case}
\usage{
NAreport(x, idvar)
}
\arguments{
\item{x}{data.frame to be analyzed}

\item{idvar}{variable to be used as id}
}
\description{
Report NA variables per case
}
\examples{
test <- data.frame(
 name =  c("john","mary","gregor"),
 x = c(1,NA,NA),
 y = c(NA,1,1),
 z = c(1,1,NA))

NAreport(test, idvar = "name")
}

