\name{pairs_with_lowess}
\alias{pairs_with_lowess}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Scatterplot Matrices with a lowess smoother
}
\description{

  A Matrix of Scatterplot with a lowess smoother is produced.
    
}

\usage{
pairs_with_lowess(data)
}

\arguments{

  \item{data}{data.frame} 

}

\value{
  Plot 
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ pairs }
\keyword{ lowess }
\keyword{ scatterplot }
\keyword{ matrix }

