% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairsWithLowess}
\alias{pairsWithLowess}
\title{Scatterplot Matrices with a lowess smoother}
\usage{
pairsWithLowess(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
Plot
}
\description{
A Matrix of Scatterplot with a lowess smoother is produced.
}
\keyword{lowess}
\keyword{matrix}
\keyword{pairs}
\keyword{scatterplot}

