% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pdf2emf}
\alias{pdf2emf}
\title{Convert a \code{pdf} to \code{emf} with \code{pstoedit} and
\code{fig2emf} (in Linux)}
\usage{
pdf2emf(files)
}
\arguments{
\item{files}{pdf files path vector}
}
\description{
Convert a \code{pdf} to \code{emf} with \code{pstoedit} and \code{fig2emf}
}

