% Generated by roxygen2 (4.0.1): do not edit by hand
\name{preprocessVarnames}
\alias{preprocessVarnames}
\title{Preprocess data.frame variable names}
\usage{
preprocessVarnames(varnames, trim = NULL)
}
\arguments{
\item{varnames}{names of a data.frame}

\item{trim}{character length of trimming. If \code{NULL}
(default) trimming is disabled.}
}
\description{
Function to preprocess variable names useful for a data.frame.
This function was created to make automatic variable name
creation from Excel files obtained by other.
}

