% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prettyPval.R
\name{prettyPval}
\alias{prettyPval}
\title{Pretty print for p-values}
\usage{
prettyPval(pvalue, space = FALSE, equal = TRUE)
}
\arguments{
\item{pvalue}{A numeric vector of p-values.}

\item{space}{Logical specifying whether a space should be inserted between
number and operator (= or <, default to FALSE).}

\item{equal}{add = where needed?}
}
\value{
The function return a string with the pretty printed p-values.
}
\description{
Pretty print for p-values.
}
\examples{
prettyPval(0.3)

pval1 <- c(3, NA, 1e-01, 1e-02, 1e-03, 1e-04, 1e-05)
prettyPval(pval1, space = TRUE)
}

