% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmAccents.R
\name{rmAccents}
\alias{rmAccents}
\title{Remove accents from a character vector}
\usage{
rmAccents(string)
}
\arguments{
\item{string}{a character vector}
}
\description{
Remove accents from a character.
}

