% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmUnprintableChars.R
\name{rmUnprintableChars}
\alias{rmUnprintableChars}
\title{Remove unprintable chars}
\usage{
rmUnprintableChars(string = NULL)
}
\arguments{
\item{string}{a character vector}
}
\description{
Remove unprintable chars
}

