% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_errors.R
\name{setup_errors}
\alias{setup_errors}
\title{Descriptive statistics for setup errors in radiotherapy}
\usage{
setup_errors(x, group)
}
\arguments{
\item{x}{vector of numeric setup error}

\item{group}{patient id vector}
}
\description{
Compute standard M, Sigma and sigma values,  according to van Herk
(2004).
}
\examples{
db <- data.frame(pt = factor(rep(1:4, each = 4)),
                           time = rep(1:4, 4),
                           y = c(
                           2,  1, 1, 1,
                           4, -2, 2, 0,
                           1, -1, 2, 2,
                           3, -3, -2, 1
                           ))
setup_errors(db$y, db$pt)
}
\references{
Marcel van Herk, Errors and Margins in Radiotherapy,
Seminars in Radiation Oncology, Vol 14, No 1 (January), 2004

Joep Stroom and Ben Heijmen, Geometrical uncertainties,
radiotherapy planning margins, and the ICRU-62 report, Radiotherapy and
Oncology 64 (2002) 75-83
}

