\name{show_pch}
\alias{show_pch}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{

  Show pch

}
\description{
  
  Show pch values.
  
}

\usage{

show_pch(extras = c("*",".","0","+","#"),
         cex = 3, col = "red3", bg = "gold", cextext = 1.2)

}

\arguments{
  \item{extras}{charachter. Further symbols to be plotted}
  \item{cex}{numeric. cex}
  \item{col}{color}
  \item{bg}{charachter. Background color}
  \item{cextext}{cex for text}
  
}

\value{
  Nothing. As a side effect the plot of pch.
}

%% \references{
%%   \href{}{}
  
%% }

\examples{
show_pch()
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ point }

