% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/showCol.R
\name{showCol}
\alias{showCol}
\title{Show R colors for graphics and grid package}
\usage{
showCol(package = c("graphics", "grid"), bg = "white", cex = 0.75,
  srt.rot = 30)
}
\arguments{
\item{package}{charachter. Which package's color to show}

\item{bg}{charachter. Background color}

\item{cex}{numeric \code{cex}}

\item{srt.rot}{degree inclination}
}
\value{
Nothing. As a side effect the plot of colors.
}
\description{
Show R colors for graphics and grid package (simple wrapper around functions
from demo(colors).
}
\examples{
showCol()
}

