% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showPch.R
\name{showPch}
\alias{showPch}
\title{Show \code{pch}}
\usage{
showPch(extras = c("*", ".", "0", "+", "#"), cex = 3, col = "red3",
  bg = "gold", cextext = 1.2)
}
\arguments{
\item{extras}{charachter. Further symbols to be plotted}

\item{cex}{numeric \code{cex}}

\item{col}{color}

\item{bg}{character. Background color}

\item{cextext}{\code{cex} for text}
}
\value{
Nothing. As a side effect the plot of \code{pch}.
}
\description{
Show \code{pch} values.
}
\examples{
showPch()
}

