% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SIR.ci}
\alias{SIR.ci}
\title{Exact CI (and estimate) for SIR/SMR}
\usage{
SIR.ci(O, E, conf.level = 0.95)
}
\arguments{
\item{O}{observed cases}

\item{E}{expected cases}

\item{conf.level}{confidence level for the returned confidence interval
(two sided)}
}
\description{
Exact CI (and estimate) for SIR/SMR
}
\examples{
## FDK Liddel, "Simple exact analysis of the standardised mortality
## ratio", Journal of Epidemiology and Community Health 1984, 38, 85-88
x <- data.frame(D = c(8,23,210), E=c(3.59,17.83,180), cl = c(0.9,0.95,0.95))
t(apply(x, 1, function(x) SIR.ci(x[1], x[2], x[3])))
}
\seealso{
Guidelines for Using
confidence intervals for public healt assessment" - Washington State
Department of Health, July 13, 2012
}

