% Generated by roxygen2 (4.0.1): do not edit by hand
\name{smartyMRSplitter}
\alias{smartyMRSplitter}
\title{Smarty Multiple Response Splitter}
\usage{
smartyMRSplitter(x, categories = NULL)
}
\arguments{
\item{x}{variable to be splitted}

\item{categories}{character vector with, if NULL all
categories given in x will be used (in alphabetic order)}
}
\description{
This function split a variable containing Smarty's output for
a multiple response question in a dummy data.frame, to be
used with \link{\code{cbind}}.
}

