\name{Table}
\alias{Table}
\title{Cross tabulation and table creation}
\usage{
Table(...)
}
\arguments{
  \item{...}{Arguments to be passed to table.}
}
\value{
The function return same results of table with NA (if
present) and margins.

%% ~Describe the value returned %% If it is a LIST, use %%
\item{comp1 }{Description of 'comp1'} %% \item{comp2
}{Description of 'comp2'}
}
\description{
This is a wrapper around which adds margins and useNA="if"
by default.
}
\keyword{Table}
\keyword{cross}
\keyword{table}
\keyword{tabulation}

