% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unoconv}
\alias{unoconv}
\title{Unoconv file converter from R}
\usage{
unoconv(..., format = stop("Format needed"))
}
\arguments{
\item{...}{file path to be exported}

\item{format}{file format}
}
\description{
Unoconv file converter from R
}

