% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%without\%}
\alias{\%without\%}
\title{Remove elements from a vector, chosing from another vector}
\usage{
x \%without\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
Return a vector x without the commont elements to another vector y.
}

