% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlxp.R
\name{xlxp}
\alias{xlxp}
\title{Faster export with openxlsx}
\usage{
xlxp(wb, sheet, x, ...)
}
\arguments{
\item{wb}{a Workbook object}

\item{sheet}{name of the sheet to be created}

\item{x}{object to be exported}

\item{...}{further arguments passed to writeData}
}
\description{
This is a wrapper around addWorksheet and writeData to allow
less typing. It export one object per sheet, the sheet will be
created accordingly to the name given
}

