% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable_descriptives.R
\name{xtable_descriptives}
\alias{xtable_descriptives}
\title{Pretty print data.frame summary in xtables}
\usage{
xtable_descriptives(x, vertical = NULL, removed = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{vertical}{character vector of db variable names to be printed
vertically. If NULL (by default) all factor are printed
vertically.}

\item{removed}{character vector of db variable names to be removed
from printing. If NULL (by default) all character variables removed.}
}
\description{
Pretty print data.frame summary in xtables
}

