\name{yapomif-package}
\alias{yapomif-package}
\alias{yapomif}
\docType{package}
\title{
  Yet another package of misc functions.
}
\description{
  
A package with misc functions
}
\details{
\tabular{ll}{
Package: \tab yapomif\cr
Type: \tab Package\cr
%% Version: \tab 0.1\cr
%% Date: \tab 2014-02-02\cr
%% Depends: \tab RODBC, survival \cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}

%% \references{
    
%% }

\keyword{ package }
\keyword{ misc }
\keyword{ code }
\keyword{ functions }

%% \seealso{
%%   \code{\link{fc.check}}
%% }
%% \examples{
%% ## TODO: make this runnable
%% ## final.fc <- ifelse(fc.check(fc), data2fc(data), fc)
%% }
