% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{yapomif}
\alias{yapomif}
\alias{yapomif-package}
\title{Yet another package of misc functions}
\description{
Yet another package of misc functions: an R package with
miscellaneous functions (wrappers for data input, data
analysis, utilities and so on).
}

