% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_grid}
\alias{add_grid}
\title{Add grid to a plot}
\usage{
add_grid(at.y, at.x, col = "lightgray", lty = "dotted", lwd =
par("lwd"))
}
\arguments{
\item{at.x}{x axis grid values}

\item{at.y}{y axis grid values}

\item{col}{grid color}

\item{lty}{line type}

\item{lwd}{line width}
}
\value{
The function adds a grid to a plot using \link{\code{abline}}.
}
\description{
Add grid to a plot using \link{\code{abline}}.
}
\keyword{add}
\keyword{grid}
\keyword{plot}

