% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Class}
\alias{Class}
\title{Extract class smartly}
\usage{
Class(x)
}
\arguments{
\item{x}{an R object (data.frame or list)}
}
\value{
A vector of classes
}
\description{
Extract class smartly from more complex data object rather than atomic
vector (eg data.frame, list) in a handy way
}
\examples{
data(airquality)
Class(airquality)
}
\keyword{class}
\keyword{typeof}

