% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Comment}
\alias{Comment}
\title{Print a message inside an Ascii border}
\usage{
Comment(msg, border = "-", wrap = 0.9 * getOption("width"))
}
\arguments{
\item{msg}{character vectors (pasted) forming the message to
be printed}

\item{border}{character used for header printing}

\item{wrap}{where to wrap}
}
\description{
Print a message inside an Ascii border; useful for working
with \code{\link{sink}}, to leave messages in the output.
}

