% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareColumns.R
\name{compareColumns}
\alias{compareColumns}
\title{Compare columns progressively}
\usage{
compareColumns(db, operator = "<", row.id = NULL, report.only = TRUE)
}
\arguments{
\item{db}{a data.frame with ordered columns}

\item{operator}{comparison operator}

\item{row.id}{an optional row id}

\item{report.only}{logical, whether to include only a report or a matrix
with all comparisons results}
}
\value{
a list with raw results and a report
}
\description{
Compare columns progressively in a dataset using a specified operator,
that tells how columns should be ordered (eg by default columns should
be increasing)
}
\examples{

compareColumns(data.frame(a = c(1,2,3), b = c(0,1,4)),
               row.id = letters[1:3])

}
