% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cramerV.R
\name{cramerV}
\alias{cramerV}
\title{Cramer's V calculator.}
\usage{
cramerV(x)
}
\arguments{
\item{x}{an R object. Currently implemented only for tables.}
}
\value{
A scalar with Cramer's V
}
\description{
Cramer's V calculator.
}
\examples{
a <- rnorm(1000) > 0.2
b <- gl(2,500)
cramerV(table(a,b))
}

