% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadb.R
\name{dadb}
\alias{dadb}
\title{Generate a data.frame of diagnostic accuracy study.}
\usage{
dadb(tn, tp, fn, fp)
}
\arguments{
\item{tn}{True negative}

\item{tp}{True positive}

\item{fn}{False negative}

\item{fp}{False positive}
}
\value{
A data.frame for diagnostic accuracy studies.
}
\description{
Generate a data.frame of diagnostic accuracy study given true negative, true
positive, false negative, false positive.
}
\examples{

db <- dadb(tn=720 , tp=190 , fn=10 , fp=80)
with(db, table(test,refstd))

}

