% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Days}
\alias{Days}
\title{Extract numeric days of month from a \code{Date},
\code{POSIXct} or \code{POSIXlt} vector.}
\usage{
Days(x)
}
\arguments{
\item{x}{Date, \code{\link{POSIXct}} or \code{\link{POSIXlt}}
vector}
}
\value{
A numeric vector.
}
\description{
Extract numeric days of month from a \code{Date},
\code{\link{POSIXct}} or \code{\link{POSIXlt}} vector.
}
\examples{
Sys.Date()
Days(Sys.Date())
Days(as.POSIXct(Sys.Date()))
Days(as.POSIXlt(Sys.Date()))
}
\keyword{Date}
\keyword{Day}
\keyword{Days}
\keyword{date}
\keyword{day}
\keyword{days}

