% Generated by roxygen2 (4.0.2): do not edit by hand
\name{days_to_weeks}
\alias{days_to_weeks}
\title{Trivial function that converts number of days in number of weeks.}
\usage{
days_to_weeks(days)
}
\arguments{
\item{days}{Numeric vector of days}
}
\value{
A numeric vector of weeks.
}
\description{
Trivial function that converts number of days in number of weeks.
}
\examples{
days <- 1:30
days_to_weeks(days)
}
\keyword{day}
\keyword{days}
\keyword{week}
\keyword{weeks}

