% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dir.clean}
\alias{dir.clean}
\title{Remove all files in a directory.}
\usage{
dir.clean(directory)
}
\arguments{
\item{directory}{Character. A path to directory to be cleaned}
}
\description{
Remove all files in a directory, by deleting the directory and creating
a new one with the same name.
}
\keyword{clean}
\keyword{directory}

