% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dir.clean}
\alias{dir.clean}
\title{Remove all files in a directory.}
\usage{
dir.clean(directory)
}
\arguments{
\item{directory}{Character. A path to directory to be cleaned}
}
\value{
Logical. Operation successful or no files in the directory.
}
\description{
Remove all files in a directory.
}
\examples{
tmp.dir <- tempdir()
file.create(paste(tmp.dir,"foo", sep = "/"))
dir.clean(tmp.dir)
dir(tmp.dir)
}
\keyword{clean}
\keyword{directory}

