% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dummy}
\alias{dummy}
\title{Generate a data.frame of dummy variables given a vector}
\usage{
dummy(x, prefix="")
}
\arguments{
\item{x}{a vector}

\item{prefix}{name prefix of returned \link{\code{data.frame}} columns}
}
\value{
a \link{\code{data.frame}} to be used with \link{\code{cbind}}
}
\description{
Generate a data.frame of dummy variables given a vector
}
\examples{
rep(1:3, 4)
dummy(rep(1:3, 4))
}

