% Generated by roxygen2 (4.0.1): do not edit by hand
\name{duplicated}
\alias{duplicated}
\title{Determine all duplicate elements}
\usage{
duplicated( x, all.dup = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector of centimeter.}

\item{all.dup}{Logical: if TRUE considers duplicated each observation
present more than one time in the oject, otherwise implement
base::duplicated algorithm (considers duplicated the second, third and so on
duplicated observation).}

\item{...}{Other arguments passed to base::duplicated.}
}
\value{
A logical vector with duplicated marked as TRUE.
}
\description{
base::duplicated determines which elements of a vector or data frame are
duplicates of elements with smaller subscripts, and returns a logical vector
indicating which elements (rows) are duplicates.
}
\details{
This wrapper allows to find all duplicated obs.
}
\keyword{duplicated}

