% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/encapsulate.R
\name{encapsulate}
\alias{encapsulate}
\title{Encapsulate text between margins.}
\usage{
encapsulate(s, start = "[", last = "]", sep = "")
}
\arguments{
\item{s}{string to be encapsulated}

\item{start}{left margin}

\item{last}{right margin}

\item{sep}{separator between string and margins}
}
\value{
An encapsulated string.
}
\description{
Encapsulate text between margins.
}
\examples{
encapsulate(letters)
encapsulate(letters, sep=" ")[1:10]
}

