\name{enSummary}
\alias{enSummary}
\title{summary with sd}
\usage{
enSummary(var)
}
\arguments{
  \item{var}{Vector to be passed to summary.}
}
\value{
The function return same results of summary adding sd.

%% ~Describe the value returned %% If it is a LIST, use %%
\item{comp1 }{Description of 'comp1'} %% \item{comp2
}{Description of 'comp2'}
}
\description{
This is a wrapper around summary which adds sd by default.
}
\keyword{summary}

