% Generated by roxygen2 (4.0.0): do not edit by hand
\name{enSummary}
\alias{enSummary}
\title{summary with sd}
\usage{
enSummary(var)
}
\arguments{
\item{var}{Vector to be passed to summary.}
}
\value{
The function return same results of summary adding sd.
}
\description{
This is a wrapper around summary which adds sd by default.
}
\keyword{summary}

