% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enSummary.R
\name{enSummary}
\alias{enSummary}
\title{Enhanced summary}
\usage{
enSummary(x, ...)
}
\arguments{
\item{x}{Vector to be passed to summary.}

\item{...}{Further arguments}
}
\description{
This is a wrapper around summary which adds sd by default.
}
\keyword{summary}

