% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAccessTable}
\alias{getAccessTable}
\title{Retrieve an MS Access (R) table from an \code{mdb} or \code{accdb} file.}
\usage{
getAccessTable(file = NULL, table = NULL, user = "admin", pw = NULL,
  ...)
}
\arguments{
\item{file}{path to the \code{mdb} or \code{accdb} file}

\item{table}{table name}

\item{user}{username (login to the file)}

\item{pw}{password (login to the file)}

\item{...}{other arguments passed to \code{\link{sqlFetch}}}
}
\value{
The function return a data.frame with data from the specified table.
}
\description{
This function is a wrapper around \code{RODBC} function to get a
table from an \code{mdb} or \code{accdb} file.
}
\examples{
\dontrun{
 getAccessTable(file="C:/prova.accdb", table="Tabella1")
}
}
\keyword{Access}
\keyword{get}
\keyword{table}

