% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getExcelSheet.R
\name{getExcelSheet}
\alias{getExcelSheet}
\title{Retrieve an MS Excel (R) sheet from an \code{xls} or \code{xlsx} file.}
\usage{
getExcelSheet(file = NULL, sheet = NULL, ...)
}
\arguments{
\item{file}{path to the \code{xls} \code{xlsx} file}

\item{sheet}{sheet name}

\item{...}{other arguments passed to \code{\link{sqlFetch}}}
}
\value{
The function return a data.frame with data from the specified sheet.
}
\description{
This function is a wrapper around \code{RODBC} function to get a
sheet from an \code{xls} or \code{xlsx} file.
}
\examples{
\dontrun{
getExcelSheet(file="C:/foo.xls", sheet="es_ps")
}
}
\keyword{Excel}
\keyword{get}
\keyword{sheet}

