% Generated by roxygen2 (4.0.0): do not edit by hand
\name{graph}
\alias{graph}
\title{Open a vectorial device file in a portable way.}
\usage{
graph(file, width, height, ...)
}
\arguments{
\item{file}{path without extension}

\item{width}{width in inches}

\item{height}{height in inches}

\item{...}{other arguments passed to \code{\link{pdf}} or
\code{\link[grDevices]{win.metafile}}}
}
\value{
Values returned by pdf or win.metafile?
}
\description{
Portable function which choose what (vectorial) device to open looking at
the system.
}
\keyword{graph}
\keyword{pdf}
\keyword{vectorial}
\keyword{win.metafile}

