% Generated by roxygen2 (4.0.1): do not edit by hand
\name{groupProgID2}
\alias{groupProgID2}
\title{Group progressive id creator from ordered id vector}
\usage{
groupProgID2(group)
}
\arguments{
\item{group}{a group vector}
}
\description{
Starting from a vector of group id (that have to be ordered),
this function creates a progressive id inside each group.
}
\examples{
set.seed(1)
## One factor id
x <- c("A", rep("C",3), rep("B",2))
cbind(x, groupProgID2(x) )
}
\keyword{group}
\keyword{id}

