% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hours.R
\name{Hours}
\alias{Hours}
\title{Extract numeric hours of month from a POSIXct or POSIXlt vector.}
\usage{
Hours(x)
}
\arguments{
\item{x}{POSIXct or POSIXlt vector}
}
\value{
A numeric vector.
}
\description{
Extract numeric hours of month from a POSIXct or POSIXlt vector.
}
\examples{

now <- Sys.time()
Hours(now)
Hours(as.POSIXlt(now))


}
