\name{inches}
\alias{inches}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Translates from cm (centimeters) to inches.
}
\description{
   
  Translates from cm (centimeters) to inches.

}

\usage{
inches(cm)
}

\arguments{
  \item{cm}{Numeric vector of centimeter.}

}

% \details{
  

% }

\value{
  A vector with inches.

%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}

}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ inches }
\keyword{ cm }

