% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inches}
\alias{inches}
\title{Translates from cm (centimeters) to inches.}
\usage{
inches(cm)
}
\arguments{
\item{cm}{Numeric vector of centimeter.}
}
\value{
A vector with inches.
}
\description{
Translates from cm (centimeters) to inches.
}
\keyword{cm}
\keyword{inches}

