% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_path}
\alias{make_path}
\title{Path-function generator.}
\usage{
make_path(path)
}
\arguments{
\item{path}{Base path of interest.}
}
\value{
A function that paste the directory set up in the call to
\code{make_path} to the argument passed in the current call.
}
\description{
This function creates another function that paste the directory set up in
the call to \code{make_path} to the argument passed in the current call,
allowing better directory parametrazion, auto-explaining code and less
typing. See example for usage.
}
\examples{
res.dir <- make_path("results")
tmp.dir <- make_path("/tmp")

## Eg for use with write.csv() or pdf()
res.dir("my_results.csv")
tmp.dir("graph.pdf")
}
\keyword{path}

