% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Minutes}
\alias{Minutes}
\title{Extract numeric minutes of month from a POSIXct or POSIXlt vector.}
\usage{
Minutes(x)
}
\arguments{
\item{x}{POSIXct or POSIXlt vector}
}
\value{
A numeric vector.
}
\description{
Extract numeric minutes of month from a POSIXct or POSIXlt vector.
}
\examples{
now <- Sys.time()
Minutes(now)
Minutes(as.POSIXlt(now))
}
\keyword{Date}
\keyword{Hour}
\keyword{Minutes}
\keyword{date}
\keyword{hour}
\keyword{minutes}

