% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Months}
\alias{Months}
\alias{Months.POSIXct}
\alias{Months.POSIXlt}
\alias{Months.default}
\title{Extract numeric months of the year from a Date, POSIXct or POSIXlt vector.}
\usage{
Months(x, string = TRUE, abbreviate = FALSE )
}
\arguments{
\item{x}{Date, POSIXct or POSIXlt vector}

\item{string}{Logical, if TRUE months string is returned, otherwise numeric}

\item{abbreviate}{logical. Should the months in string format be
abbreviated?}
}
\value{
A vector with extracted months.
}
\description{
Extract numeric months of the year from a Date, POSIXct or POSIXlt vector.
}
\examples{
Sys.Date()
Months(Sys.Date())
Months(Sys.Date(), string = FALSE)
Months(Sys.Date(), string = TRUE, abbreviate = TRUE)
## from base package
Months(Sys.Date())
Months(as.POSIXct(Sys.Date()))
Months(as.POSIXlt(Sys.Date()))
}
\keyword{Date}
\keyword{Month}
\keyword{Months}
\keyword{date}
\keyword{month}
\keyword{months}

