% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAreport.R
\name{NAreport}
\alias{NAreport}
\title{Report NA variables per case}
\usage{
NAreport(x, idvar = NULL)
}
\arguments{
\item{x}{data.frame to be analyzed}

\item{idvar}{variable to be used as id}
}
\description{
Report NA variables per case
}
\examples{
test <- data.frame(
 name =  c("john","mary","gregor"),
 x = c(1,NA,NA),
 y = c(NA,1,1),
 z = c(1,1,NA))
test2 <- test
test2$surname <- c("doe", "foo", "bar")
test2 <- test2[c("name","surname","x","y","z")]

test
NAreport(test, "name")

test2
NAreport(test2, c("name", "surname"))
}

