% Generated by roxygen2 (4.0.1): do not edit by hand
\name{open_xl}
\alias{open_xl}
\title{Open an xlsx file}
\usage{
open_xl(file=NULL)
}
\arguments{
\item{file}{path to the Excel (xls/xlsx) file}
}
\description{
This function tries to open an xlsx file with the
proper application, in a portable manner.
}
\examples{
require(openxlsx)
example(writeData)
open_xl("writeDataExample.xlsx")
}
\author{
Luca Braglia
}

