% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairs_with_lowess}
\alias{pairs_with_lowess}
\title{Scatterplot Matrices with a lowess smoother}
\usage{
pairs_with_lowess(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
Plot
}
\description{
A Matrix of Scatterplot with a lowess smoother is produced.
}
\keyword{lowess}
\keyword{matrix}
\keyword{pairs}
\keyword{scatterplot}

