% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsWithLowess.R
\name{pairsWithLowess}
\alias{pairsWithLowess}
\title{Scatterplot Matrices with a lowess smoother}
\usage{
pairsWithLowess(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
Plot
}
\description{
A Matrix of Scatterplot with a lowess smoother is produced.
}
