% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgDownloads}
\alias{pkgDownloads}
\title{Retrieve package download logs from cran-logs.r-studio.com}
\usage{
pkgDownloads(from = Sys.Date() - 1, to = Sys.Date() - 1)
}
\arguments{
\item{from}{starting \code{Date} (or something coercible to)}

\item{to}{ending \code{Date} (or something coercible to)}
}
\description{
Retrieve package download logs from cran-logs.r-studio.com
}
\examples{
\dontrun{
## looking at yesterday's data for aplore3
db <- pkgDownloads()
barplot(with(db, tapply(package \%in\% "aplore3", date, sum )), las = 2)
}
}

