% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgDownloads}
\alias{pkgDownloads}
\title{Retrieve package download logs from cran-logs.r-studio.com}
\usage{
pkgDownloads(from = Sys.Date() - 1, to = Sys.Date() - 1)
}
\arguments{
\item{from}{starting \code{Date} (or something coercible to)}

\item{to}{ending \code{Date} (or something coercible to)}
}
\description{
Retrieve package download logs from cran-logs.r-studio.com
}
\examples{
\dontrun{
## looking at yesterday's data for aplore3 and aprean3
db <- pkgDownloads(from = Sys.Date()-4)
dbSel <- db[db$package \%in\% c("aplore3", "aprean3"), ]
res <- with(dbSel, table(date, package))
barplot(res, beside = TRUE)
}
}

