\name{pretty_pval}
\alias{pretty_pval}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Pretty print for p-values
}
\description{
  
  Pretty print for p-values.
  
}

\usage{
pretty_pval(pvalue, space)
}

\arguments{
  \item{pvalue}{A numeric vector of p-values.}

  \item{space}{Logical specifying whether a space shoud be inserted
    between number and operator (= or <, default to FALSE).}
}

% \details{
  

% }

\value{
  The function return a string with the pretty printed p-values.

%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}

}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ pretty }
\keyword{ print }
\keyword{ p-value }
\keyword{ p-values }
\keyword{ p }
\keyword{ value }
\keyword{ values }

