% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quickROC.R
\name{quickROC}
\alias{quickROC}
\title{A quick ROC plot.}
\usage{
quickROC(test, refstd, direction = c("auto", "<", ">"), plot.roc = TRUE,
  main = "")
}
\arguments{
\item{test}{Marker}

\item{refstd}{Reference standard (binary variable)}

\item{direction}{Charachter. Direction passed to \code{pROC::roc}}

\item{plot.roc}{Logical. Plot ROC curve?}

\item{main}{Graph main title}
}
\value{
The function plot the graph and return a list with ROC statistics
}
\description{
A quick ROC plot. It's a simple wrapper around \code{pROC::roc}
}

