\name{recode}
\alias{recode}
\title{Vector recode utility}
\usage{
recode(from=NULL, rec.direct=NULL)
}
\arguments{
  \item{from}{Vector to be recoded.}

  \item{rec.direct}{matrix containing 2 columns; first is
  "from", second is "to".}
}
\value{
A vector with new codes (if/where modified).
}
\description{
Yet another recode utility
}
\keyword{recode}
\keyword{replace}

