\name{recode}
\alias{recode}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Vector recode utility
}
\description{
  
  Yet another recode utility
  
}

\usage{
recode(from=NULL, rec.direct=NULL)
}

\arguments{
  \item{from}{Vector to be recoded.}
  \item{rec.direct}{matrix containing 2 columns; first is "from", second
  is "to".}
}

% \details{
  

% }

\value{
  A vector with new codes (if/where modified).

%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}

}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ recode }
\keyword{ replace }

