% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmSpaces.R
\name{rmSpaces}
\alias{rmSpaces}
\title{Remove spaces from a character vector}
\usage{
rmSpaces(string)
}
\arguments{
\item{string}{a character vector}
}
\description{
This function removes spaces (all leading and trailing ones,
and unique in-between) from a character vector.
}
\examples{
test <- c("  test ", "  mediterranean  sea  ")
rmSpaces(test)
}

