% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_pch}
\alias{show_pch}
\title{Show pch}
\usage{
show_pch(extras = c("*",".","0","+","#"), cex = 3, col = "red3", bg =
"gold", cextext = 1.2)
}
\arguments{
\item{extras}{charachter. Further symbols to be plotted}

\item{cex}{numeric. cex}

\item{col}{color}

\item{bg}{charachter. Background color}

\item{cextext}{cex for text}
}
\value{
Nothing. As a side effect the plot of pch.
}
\description{
Show pch values.
}
\examples{
show_pch()
}
\keyword{point}

