\name{Table}
\alias{Table}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Cross tabulation and table creation
}
\description{
  
  This is a wrapper around which adds margins and useNA="if" by default.
  
}

\usage{
Table(...)
}

\arguments{
  \item{...}{Arguments to be passed to table.}
}

% \details{
  

% }

\value{
  The function return same results of table with NA (if present) and margins.

%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}

}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ table }
\keyword{ Table }
\keyword{ cross }
\keyword{ tabulation }

