% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Table}
\alias{Table}
\title{Cross tabulation and table creation}
\usage{
Table(...)
}
\arguments{
\item{...}{Arguments to be passed to table.}
}
\value{
The function return same results of table with NA (if present) and
margins.
}
\description{
This is a wrapper around which adds margins and useNA="if" by default.
}
\keyword{Table}
\keyword{cross}
\keyword{table}
\keyword{tabulation}

