% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tteep}
\alias{tteep}
\title{Common time to event end-points calculator.}
\usage{
tteep(prog =!is.na(prog.date),
prog.date=NULL,
death = !is.na(death.date),
death.date=NULL,
start.date=NULL,
last.fup=NULL,
OS=TRUE,
PFS=TRUE,
TTP=TRUE,
visual.check=FALSE)
}
\arguments{
\item{prog}{Progression Indicator}

\item{prog.date}{Progression Date}

\item{death}{Death indicator}

\item{death.date}{Death date}

\item{start.date}{Start date}

\item{last.fup}{Last fup}

\item{OS}{Calculate Overall Survival?}

\item{PFS}{Calculate Progression Free Survival?}

\item{TTP}{Calculate Time To Progression?}

\item{visual.check}{View results}
}
\value{
A data frame to be used with \code{\link{cbind}}.
}
\description{
This function calculates common oncology time to event end-points (Overall
survival, Progression free survival, time to relapse).
}
\keyword{death}
\keyword{disease-free}
\keyword{end-point}
\keyword{event}
\keyword{oncology}
\keyword{outcome}
\keyword{overall}
\keyword{progression}
\keyword{progression-free}
\keyword{survival}
\keyword{time}
\keyword{to}

