% Generated by roxygen2 (4.0.2): do not edit by hand
\name{univTable}
\alias{univTable}
\title{Univariate table with absolute, relative, cumulative frequencies.}
\usage{
univTable(var=NULL, totals=TRUE, useNA="ifany",
NA.string = "NA", round.digits=3, sorting=NULL)
}
\arguments{
\item{var}{vector}

\item{totals}{print totals?}

\item{useNA}{print NA?}

\item{NA.string}{character used for NA's columns title}

\item{round.digits}{number of rounding digits}

\item{sorting}{sorting can be "\code{asc} or "\code{desc}"}
}
\value{
A matrix with frequencies.
}
\description{
Univariate table with absolute, relative, cumulative frequencies.
}
\keyword{frequencies}
\keyword{table}
\keyword{univariate}

