% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xtile}
\alias{xtile}
\title{Cut numeric x in quantile}
\usage{
xtile(x,n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of quantiles}
}
\value{
A factor vector.
}
\description{
Cut numeric x in quantile
}
\examples{
tertiles <- xtile(rnorm(1000),3)
table(tertiles)
}
\keyword{cut}
\keyword{quantiles}
\keyword{xtile}

