% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Years.R
\name{Years}
\alias{Years}
\title{Extract numeric years from a Date, POSIXct or POSIXlt vector.}
\usage{
Years(x)
}
\arguments{
\item{x}{Date, POSIXct or POSIXlt vector}
}
\value{
A numeric vector.
}
\description{
Extract numeric years from a Date, POSIXct or POSIXlt vector.
}
\examples{
Sys.Date()
Years(Sys.Date())
Years(as.POSIXct(Sys.Date()))
Years(as.POSIXlt(Sys.Date()))
}

